Fixes trac:4627: Move r31759 to v1.8 branch (Close the incoming socket if)
---svn-pre-commit-ignore-below---

r31759 [[BR]]
Close the incoming socket if we aren't going to accept it

cmr=v1.8.2:reviewer=rhc


r31763 [[BR]]
Just pick a module to handle the incoming connection if no direct interface is identified. Siegmar hit it because his IP/netmask is disjoint, but a router was able to make the connection.

Refs trac:4627

This commit was SVN r31789.

The following SVN revision numbers were found above:
  r31759 --> open-mpi/ompi@e605e7337960eb9c044868e2f34f3dbc67369f35
  r31763 --> open-mpi/ompi@ad0e8f841dd496f45dad55e61ec711759e9776e8

The following Trac tickets were found above:
  Ticket 4627 --> https://svn.open-mpi.org/trac/ompi/ticket/4627