Fixes trac:4647: Move r31806, r31841 to v1.8 branch (pml ob1: use the pre-computed)
---svn-pre-commit-ignore-below---

r31806 [[BR]]
pml_ob1: ensure to have enough space for send/recvreq on stack

r30343 introduced the optimization of putting the OB1 sendreq and
recvreq on the stack for blocking sends and receives.  However, the
requests did not contain enough storage for the data that is normally
immediately ''after'' the request (e.g., BTL data).

This commit changes these requests to be pointers and to use alloca()
to get enough total space for the OB1 request and all the associated
data.

The change is smaller than it looks; most of it is just changing from
"foo.bar" to "foo->bar" notation (etc.).

Submitted by Jeff, reviewed by Nathan.  But we want George to look at
this (and get a little soak time on the trunk) before moving to v1.8.

cmr=v1.8.2:reviewer=bosilca

r31841 [[BR]]
pml ob1: use the pre-computed size from the free lists

Based on a suggestion from George on #31806, use the pre-computed
sizes rather than duplicating the computation math (which may change
someday in the future).

cmr=v1.8.2:ticket=trac:4647

This commit was SVN r31847.

The following SVN revision numbers were found above:
  r30343 --> open-mpi/ompi@2b57f4227ebfcfd5f86fdb6ac37c6bb81cecf521
  r31806 --> open-mpi/ompi@025e4a852bf0af03361bf9b64f22d57951c8479f
  r31841 --> open-mpi/ompi@b0a6e42f45972edb23de7cf78d32117298894da0

The following Trac tickets were found above:
  Ticket 4647 --> https://svn.open-mpi.org/trac/ompi/ticket/4647