Fixes trac:4699: Move r31959 to v1.8 branch (Correct the bind-in-place algorithm to)
---svn-pre-commit-ignore-below---

r31959 [[BR]]
Correct the bind-in-place algorithm to better handle comm_spawn. If the location identified by the mapper is already occupied by procs from another job, then we need to shift either right or left until we find an unoccupied location where we can be bound. If nothing is available, then check for the overload flag (and bind us in the original location if provided), or see if this was the default binding policy instead of one specified by the user - if so, then just don't bind this process.

cmr=v1.8.2:reviewer=rhc

This commit was SVN r31993.

The following SVN revision numbers were found above:
  r31959 --> open-mpi/ompi@638c24f655139615add842fd1036f1460af5871a

The following Trac tickets were found above:
  Ticket 4699 --> https://svn.open-mpi.org/trac/ompi/ticket/4699