osc/rdma: ensure eager sends are active before checking for sync errors in self optimization
This addresses an issue found with the MPICH pscw_ordering test. Eager sends
were not yet active (which is ok for the standard path) but not ok for the
self optimization. Fixed by waiting for all post messages before checking
the sync state.

Fixes #4724

Tracking the 1.8.2 issue in this CMR.

cmr=v1.8.2:reviewer=bbenton