Fixes trac:4436: Move r31189, r31193 to v1.8 branch (coll/ml: addendum to r31189. increment)
---svn-pre-commit-ignore-below---

r31189 [[BR]]
coll/ml: improve the buffer size calculation and ensure the bcol_index in
a hierarchy actually matches a bcol that is in use.

There was a bug in one of the paths to calculate the ml buffer size. I fixed
the bug and squashed all the paths together to avoid further issues (the
result was correct in another path that calculated the same value).

Additionally, the i_hier was being used as the bcol_index. This is not
correct in a couple of cases so I added a variable to keep track of the
real bcol_index.

cmr=v1.8:reviewer=pasha

r31193 [[BR]]
coll/ml: addendum to r31189. increment the bcol_index

cmr=v1.8:ticket=trac:4436

This commit was SVN r32114.

The following SVN revision numbers were found above:
  r31189 --> open-mpi/ompi@c7d830f4b9ba7dd22cf92f6fa0f4d2d823621916
  r31193 --> open-mpi/ompi@15a8c9d7b807777f3e531b04b861feab78b387d1

The following Trac tickets were found above:
  Ticket 4436 --> https://svn.open-mpi.org/trac/ompi/ticket/4436