Fixes trac:4564: Move r31522 to v1.8 branch (btl/vader: improve performance under heavy)
---svn-pre-commit-ignore-below---

r31522 [[BR]]
btl/vader: improve performance under heavy load and eliminate a racy
feature

This commit should fix a hang seen when running some of the one-sided
tests. The downside of this fix is it reduces the maximum size of the
messages that use the fast boxes. I will fix this in a later commit.

To improve performance under a heavy load I introduced sequencing to
ensure messages are given to the pml in order. I have seen little-no
impact on the message rate or latency with this change and there is a
clear improvement to the heavy message rate case.

Lets let this sit in the trunk for a couple of days to ensure that
everything is working correctly.

cmr=v1.8.2:reviewer=jsquyres

This commit was SVN r32167.

The following SVN revision numbers were found above:
  r31522 --> open-mpi/ompi@0849d61e381c7b12cc992836fa66486d7108b17c

The following Trac tickets were found above:
  Ticket 4564 --> https://svn.open-mpi.org/trac/ompi/ticket/4564