Fixes trac:4774: Move r32160 to v1.8 branch (osc/rdma: check for more types)
---svn-pre-commit-ignore-below---

r32160 [[BR]]
osc/rdma: check for more types of window access violations

This commit adds a check to see if the target is in an access epoch. If
not we return OMPI_ERR_RMA_SYNC. This fixes test_start3 in the onesided
test suite. The cost of this extra check is 1 byte/peer for the boolean
flag indicating that the peer is in an access epoch.

I also fixed a problem where mupliple unexpected post messages are not
correctly handled.

cmr=v1.8.2:reviewer=jsquyres

This commit was SVN r32171.

The following SVN revision numbers were found above:
  r32160 --> open-mpi/ompi@b6abe689720232a987e302889f5b30808d636e56

The following Trac tickets were found above:
  Ticket 4774 --> https://svn.open-mpi.org/trac/ompi/ticket/4774