Fixes trac:4787: Move r32229 to v1.8 branch (Get the blasted singleton comm_spawn)
---svn-pre-commit-ignore-below---

r32229 [[BR]]
Get the blasted singleton comm_spawn working again. There remain problems with the Slurm interaction in this use-case as the PMI components (if configured to build) try to run even when a Slurm allocation hasn't been made, but I leave that to someone else to resolve. I did, however, tell the Slurm ess to quit interfering with applications launched in this use-case by ORTE daemons, so things do work when inside a Slurm allocation.

Also discovered that the rsh launcher is not picking up --enable-orterun-prefix-by-default when invoked during singleton comm_spawn, but I was unable to see why that was happening and ran out of time.

cmr=v1.8.2:reviewer=rhc

This commit was SVN r32279.

The following SVN revision numbers were found above:
  r32229 --> open-mpi/ompi@1feaffbb159032fd691d42d326da6fbd413019eb

The following Trac tickets were found above:
  Ticket 4787 --> https://svn.open-mpi.org/trac/ompi/ticket/4787