Fixes trac:4810: Move r32309 to v1.8 branch (usnic: check connectivity on first)
---svn-pre-commit-ignore-below---

r32309 [[BR]]
usnic: check connectivity on first communication to a peer

Previously, we were only checking connectivity upon first ''send'' to
a peer.  But this ignores the case where the first communication to a
peer is actually an ACK -- i.e., we successfully received something
from the peer and we need to send an ACK back.  So we need to verify
that the ACK will actually get there.

Specifically, certain asymmetric routing cases can lead to a hang if
we don't check the connectivity in both directions.  E.g., if the
sender is able to get traffic to the receiver, but the receiver is
unable to get traffic back to the sender because it made a different
routing decision than the sender.

In this case, the connectivity checker from the sender could succeed
(because the connectivity checker will ACK along the same path in
which the ping was received), but sending a BTL ACK could fail
(because the BTL ACK will be sent back along the path chosen by the
graph algorithm, which, in an erroneous asymmetric routing scenario,
may be different/wrong).

Hence, we want to trigger the connectivity checker at the first
communication from A->B, which may either be a BTL send or an ACK.

Reviewed by Dave Goodell.

cmr=v1.8.2:reviewer=ompi-rm1.8

This commit was SVN r32311.

The following SVN revision numbers were found above:
  r32309 --> open-mpi/ompi@6ae45b34fc5b7ce611a4390e1b6cfc3687804be8

The following Trac tickets were found above:
  Ticket 4810 --> https://svn.open-mpi.org/trac/ompi/ticket/4810