Fixes trac:4817: Add --with-usnic option to usnic BTL configure.m4
---svn-pre-commit-ignore-below---

Custom patch on ticket based on:

r32349 [[BR]]
usnic: add --with-usnic configure switch
    
If --with-usnic is specified and we can't build the usnic BTL, abort.
If --without-usnic is specified, gracefully skip building the usnic
BTL.  If neither is specified, do the OMPI-default behavior: try to
configure/build the usnic BTL, and if we can't, skip it.
    
Fixes CSCuq13889.

This commit was SVN r32374.

The following SVN revision numbers were found above:
  r32349 --> open-mpi/ompi@10a58992af6b2109e4690ef4bd76b620ee96f905

The following Trac tickets were found above:
  Ticket 4817 --> https://svn.open-mpi.org/trac/ompi/ticket/4817