Fixes trac:4824: Add usnic connectivity checker source and expected ping checks
---svn-pre-commit-ignore-below---

Custom patch on ticket based on:

r32383 [[BR]]
usnic: cagent now checks that incoming pings are expected

Previously, the connectivity agent was pretty dumb: it took whatever
pings it got and ACKed them.  Then we added an agent check to ensured
that the ping actually came from the source interface that it said it
came from.  Now we add another check such that when a ping is received
on interface X that corresponds to usnic module Y, we ensure that the
source interface of the ping is on the all_endpoints list for module Y
(i.e., module Y expects to be able to talk to that peer interface).

This detects cases where peers have come to different conclusions
about which interfaces should be used to communicate (which is bad!).
This usually reflects a network misconfiguration.

Fixes CSCuq05389.

This commit was SVN r32387.

The following SVN revision numbers were found above:
  r32383 --> open-mpi/ompi@ff4717b7273140c1f7f3d7d8d2ef2eae377acbdd

The following Trac tickets were found above:
  Ticket 4824 --> https://svn.open-mpi.org/trac/ompi/ticket/4824