usnic: update help messages
These messages were committed in the v1.8 branch in r32341, but were
never committed to the trunk (because we were waiting for the OPAL BTL
move).  This commit brings the trunk and v1.8 help messages in line
with each other.

This commit was SVN r32445.

The following SVN revision numbers were found above:
  r32341 --> open-mpi/ompi@5e752b4abaa4859a79372dbb435aceceb17a9eba