check-help-strings.pl: work better when running in a subdir
The script used to get confused if you ran it in a subdir of the OMPI
tree (e.g., it would report help files not found if source code
referred to a help file that was not under the subdir in which it was
run).

Now the script searches the whole tree for help files, but only
analyzes the source from the subdir in which it was started.  Meaning:
it should always find all help files, and won't get confused like it
did before.

This commit was SVN r32548.