Fixes trac:4866: Move r32546 to v1.8 branch (usnic: ensure to have a)
Hand-edited to port from trunk

---svn-pre-commit-ignore-below---

r32546 [[BR]]
usnic: ensure to have a safe destruction of an opal_list_item_t

It turns out that we ''can'' get to the endpoint destructor with the
endpoint still on the "endpoints needing ACKs" list.  So if it's on
the list, remove it first, and then DESTRUCT the opal_list_item_t.

This prevents an assert() fail in debug builds.  We'd like to let this
soak over the weekend.

cmr=v1.8.2:reviewer=dgoodell

This commit was SVN r32560.

The following SVN revision numbers were found above:
  r32546 --> open-mpi/ompi@ac7c907f8d0fc79f15fedcf9fabc41d29255618a

The following Trac tickets were found above:
  Ticket 4866 --> https://svn.open-mpi.org/trac/ompi/ticket/4866