Fixes trac:4870: Move r32553 to v1.8 branch (If an initial address fails)
---svn-pre-commit-ignore-below---

r32553 [[BR]]
If an initial address fails to connect, record that fact and attempt the next address for that proc. If nothing succeeds, then declare failure.

cmr=v1.8.2:reviewer=edgar


r32554 [[BR]]
Frazzle-frump! Missed that we reset the peer state just before the new check.

Refs trac:4870


r32556 [[BR]]
Track down the last piece of the connection problem. It appears that
providing a netmask of 0 to opal_net_samenetwork results in everything
looking like it is on the same network. Hence, we were not retaining any
of the alternative addresses, so we had no other way to check them.

Refs trac:4870


r32558 [[BR]]
Remove an unnecessary optimization that can cause more trouble than it's worth - just try all the addresses that are given to us.

Refs trac:4870

This commit was SVN r32563.

The following SVN revision numbers were found above:
  r32553 --> open-mpi/ompi@0a91fdf85f0808d93e1d8750225dc2a8e4fe4220
  r32554 --> open-mpi/ompi@343038af7b948acd025aa9413688d000f14cde44
  r32556 --> open-mpi/ompi@fa28710d534ea27ac68343719f1ab3f78ce4addd
  r32558 --> open-mpi/ompi@b4511913f6068cf7291247aca92cb0ee1543e94e

The following Trac tickets were found above:
  Ticket 4870 --> https://svn.open-mpi.org/trac/ompi/ticket/4870