Fixes trac:4660: Move r31828 to v1.8 branch (pml/ob1: as per past RFC)
---svn-pre-commit-ignore-below---

r31828 [[BR]]
pml/ob1: as per past RFC bring the inline send optimization to
MPI_Isend.

I filed an RFC for this optimization some time back. It is a
relatively simple optimization. If the data associated with an
MPI_Isend can be put on the wire without allocating an MPI_Request
then do so. In this case we can legally return omp_request_empty
which will correctly indicate that the request is complete and that is
was not cancelled (these are the only requirements on send requests).

cmr=v1.8.3:reviewer=bosilca

This commit was SVN r32610.

The following SVN revision numbers were found above:
  r31828 --> open-mpi/ompi@a1d5ce0893df6b42470221744323429a3ff610cc

The following Trac tickets were found above:
  Ticket 4660 --> https://svn.open-mpi.org/trac/ompi/ticket/4660