make_dist_tarball: check for a dirty source tree
Have the make_dist_tarball script check to ensure that the source tree
is clean before continuing.  This ensures that we don't accidentally
build a distribution tarball with something that is not committed in
the repo.

There is a --dirtyok option to override this check, and if you access
this script via the "make_tarball" link, --dirtyok is added to the
default set of options.

cmr=v1.8.3:reviewer=rhc

This commit was SVN r32623.