Fixes trac:4883: Move r32623 to v1.8 branch (make_dist_tarball: check for a dirty)
---svn-pre-commit-ignore-below---

r32623 [[BR]]
make_dist_tarball: check for a dirty source tree

Have the make_dist_tarball script check to ensure that the source tree
is clean before continuing.  This ensures that we don't accidentally
build a distribution tarball with something that is not committed in
the repo.

There is a --dirtyok option to override this check, and if you access
this script via the "make_tarball" link, --dirtyok is added to the
default set of options.

cmr=v1.8.3:reviewer=rhc

This commit was SVN r32695.

The following SVN revision numbers were found above:
  r32623 --> open-mpi/ompi@008454a4a888be1eb8a58efaae1b03094618b30d

The following Trac tickets were found above:
  Ticket 4883 --> https://svn.open-mpi.org/trac/ompi/ticket/4883