Fixes trac:4897: Move r32686 to v1.8 branch (Wrap MCA params)
---svn-pre-commit-ignore-below---

r32686 [[BR]]
Since we've run into the situation where someone puts a script wrapper around a launcher such as srun, we need to always protect MCA cmd line params with quotes. This means we also need to protect the backend from quotes coming into the system as part of a value, or else the parser gets confused.

So add a new function for wrapping MCA arguments, and tell the backend parser to ignore/remove leading/trailing quotes.

cmr=v1.8.3:reviewer=jsquyres


r32719 [[BR]]
Check for multiple declarations of a given MCA param and error out if detected as that can create an ambiguous definition of the param value.

Refs trac:4897


r32726 [[BR]]
plm_rsh_module.c: output to the framework output

Trivial fix from r32686: don't output to stream 0, but rather to
orte_plm_base_framework.framework_output (this is the way it was
before r32686).  In reality, this is going to end up being stream 0,
anyway, but we might as well be pedantically correct...

Refs trac:4897.

r32731 [[BR]]
Not sure how we could get to this point without having already detected the error, but just to be safe - check for end-of-array and return if error.

Refs trac:4897

This commit was SVN r32732.

The following SVN revision numbers were found above:
  r32686 --> open-mpi/ompi@4df1aa63f781c709c136e9f94b5544974f28477a
  r32719 --> open-mpi/ompi@0445052a1caae5c480a76983ac334387a9e47b16
  r32726 --> open-mpi/ompi@e95ed94a947f68ca2f6e909611166cda9fc229e3
  r32731 --> open-mpi/ompi@fad43844635a991d144dfcec436763a13e7b62c9

The following Trac tickets were found above:
  Ticket 4897 --> https://svn.open-mpi.org/trac/ompi/ticket/4897