Fixes trac:4908: Move r32743 to v1.8 branch (Provide a better help message)
---svn-pre-commit-ignore-below---

r32743 [[BR]]
Provide a better help message when we are unable to complete a connection due to a firewall.

cmr=v1.8.3:reviewer=jsquyres


r32746 [[BR]]
Can't use show_help as the RML has already been enabled, but we haven't successfully connected back to the HNP. So use opal_output instead and hardwire the message.

Refs trac:4908


r32748 [[BR]]
Try to provide a real hostname for the remote host to aid in debugging

Refs trac:4908

This commit was SVN r32756.

The following SVN revision numbers were found above:
  r32743 --> open-mpi/ompi@a74428513d5b010c49f4b3fe1ec3c27e63c9b0d8
  r32746 --> open-mpi/ompi@09aecea55afd525c03193566eedd7dcfbf0c9672
  r32748 --> open-mpi/ompi@414f4e97837f9709831da17135ef9e9580035be8

The following Trac tickets were found above:
  Ticket 4908 --> https://svn.open-mpi.org/trac/ompi/ticket/4908