Fixes trac:4909: Move r32744 to v1.8 branch (Try to resolve the simultaneous)
---svn-pre-commit-ignore-below---

r32744 [[BR]]
Try to resolve the simultaneous connection problem by being a little more careful about the choice of returned status when a connection is refused. As before, have the higher vpid of the two peers retry the connection, while the lower one waits. This can happen in a couple of places, so try to hit them all. Since this is hard to test, will ask Gilles to give it a try since he's the one who is seeing it.

cmr=v1.8.3:reviewer=rhc


r32754 [[BR]]
oob/tcp: fix a race condition

Refs trac:4909

This commit was SVN r32757.

The following SVN revision numbers were found above:
  r32744 --> open-mpi/ompi@4bbc9a28d69b8707d5f160b96335837019a4a137
  r32754 --> open-mpi/ompi@5fa2b6c59c7d2b754c4b20e88150aab16abd16cd

The following Trac tickets were found above:
  Ticket 4909 --> https://svn.open-mpi.org/trac/ompi/ticket/4909