Fixes trac:4914: Move r32752 to v1.8 branch (fortran: configure test for "optional")
---svn-pre-commit-ignore-below---

r32752 [[BR]]
fortran: configure test for "optional" keyword was too aggressive

After discussions with Craig, it looks like the check in this
configure test was too aggressive and based on a prior mpi_f08
implementation model.  We don't use "value" any more, and we also
don't pass optional parameters back to C code, so this test was
checking far more than it needed to... in a non-standard way (which
breaks in the Intel 2015 Fortran compiler).  All we ''really'' need is
to check whether the compiler supports the "optional" keyword.  So
make the test ''much'' simpler and just check whether ''optional''
compiles successfull or not.

Reviewed by Craig Rasmussen

cmr=v1.8.3:reviewer=ompi-rm1.8

This commit was SVN r32762.

The following SVN revision numbers were found above:
  r32752 --> open-mpi/ompi@cf56a369c223a103e4c9508bcdc1daca5cb196bf

The following Trac tickets were found above:
  Ticket 4914 --> https://svn.open-mpi.org/trac/ompi/ticket/4914