Fixes trac:4916: Move MCA base var alignment fixes to v1.8
---svn-pre-commit-ignore-below---

r32716 [[BR]]
mca_base_var: fix possibilty of unaligned variable assignments

Add a debugging check that ensures that the registered storage is
aligned appropriately for the type that is specified.

When we know that the storage is properly aligned, we can cast the
mbv_storage to the appropriate type and then simply do the assignment.
We used to do this assignment via a union, but clang's
-fsanitizer=alignment complained about this.

This commit was SVN r32763.

The following SVN revision numbers were found above:
  r32716 --> open-mpi/ompi@d244b7b860634147a41400e4488db5e51054d7f2

The following Trac tickets were found above:
  Ticket 4916 --> https://svn.open-mpi.org/trac/ompi/ticket/4916